+++ Test/function_test.py
@@ -190,7 +190,7 @@ Example : online_help("J") or online_help(2)
 >>> from pycm.overall_funcs import *
 >>> from pycm.pycm_class_func import *
 >>> from pycm.pycm_ci import *
->>> from pycm.pycm_interpret import *
+>>> from pycm.interpret import *
 >>> from pycm.utils import *
 >>> color_check("red")
 [255, 0, 0]